+++ Something/Startup.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Bookshelves
+namespace Something
 {
     public class Startup
     {
